Introduce an interface to describe the data structure describing a cache in the build operation result, to reduce coupling to the impl.
